requirements.txt
# torch/torchaudio installed via torch.js
gradio>=4.0.0
yt-dlp>=2024.0.0
git+https://github.com/m-bain/whisperX.git
transformers>=4.35.0
sentencepiece
protobuf
demucs>=4.0.0
pydub>=0.25.1
soundfile>=0.12.0
librosa>=0.10.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
accelerate>=0.25.0
bitsandbytes>=0.43.0
# qwen-tts installed separately via tts_install.js (avoids dependency conflicts)
